Обновление конфигурации сервисов и рефакторинг кода
Изменения в `docker-compose.yml` включают обновление конфигурации сервисов, добавление новых переменных окружения и зависимостей, а также изменение условий для сервисов.

В классе `MerchandiseRepository` изменен метод `CreateAsync` для возврата идентификатора созданного объекта, добавлен метод `UpdateAsync`.

В классе `SkuPresetRepository` метод `SaveAsync` заменен на `CreateAsync`.

Добавлено сохранение изменений в базе данных в `DomainEventPublishingBehavior` перед публикацией событий.

В `GiveOutMerchandiseCommandHandler` добавлено создание `skuPreset` перед обновлением статуса заявки.

Убраны зависимости от `IEmailService` в `MerchandiseRequestDeclinedHandler` и `MerchandiseRequestGiveOutHandler`.

Изменены имена столбцов в конфигурациях `MerchandiseRequestConfiguration` и `SkuPresetConfiguration`.

В миграциях базы данных обновлены имена таблиц и добавлены новые миграции для создания таблиц `merchandise_requests`, `sku_presets` и `sku_preset_skus`.

Изменена логика получения адреса сервиса в `HostBuilderExtensions.cs`.

Добавлен вызов метода `AddStockGrpcServiceClient` в `Program.cs` для регистрации клиента gRPC.

Обновлен адрес сервера в `appsettings.json` для `StockApiGrpcServiceConfiguration`.

+++ src/OzonEdu.MerchandiseService.Migrator/Migrations/0011_SkuPresetsTable.cs
@@ -9,7 +9,7 @@ namespace OzonEdu.MerchandiseService.Migrator.Migrations
 		{
 			Create
 				.Table("sku_presets")
-				.WithColumn("id").AsInt32().PrimaryKey().Identity()
+				.WithColumn("id").AsInt64().PrimaryKey().Identity()
 				.WithColumn("preset_type_id").AsInt32().NotNullable();
 		}
